🐛 fix: restore message action bar on tap after deletion
